+++ messaging-core/src/main/java/com/nabla/sdk/messaging/core/domain/entity/Conversation.kt
@@ -48,6 +48,7 @@ public data class Conversation(
     val patientUnreadMessageCount: Int,
     val providersInConversation: List<ProviderInConversation>,
     val pictureUrl: EphemeralUrl?,
+    val isLocked: Boolean,
 ) {
     public companion object
 }